% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_structure.R
\name{fit_structure}
\alias{fit_structure}
\title{Fit a population structure to data.}
\usage{
fit_structure(
  samples,
  ages = NA,
  sizes = NA,
  weights = NA,
  initpars,
  model = "BOTH",
  feedinggroups = NA,
  feedingbounds = NA,
  logscale = TRUE,
  shrink = TRUE,
  opts = list(algorithm = "NLOPT_LN_SBPLX", xtolrel = 1e-06, maxeval = 1e+06)
)
}
\arguments{
\item{samples}{A discrete (integer) value indicating the timepoint of the observation. Timepoints should have a constant interval and consecutive timepoints should have consecutive values. Or a dataset with 3 or 4 colums containing timepoints, ages, sizes and (optionally) weights as described.}

\item{ages}{A discrete (integer) value indicating the age of the observed individual. Ages should be on the same timescale as the timepoints. Unused if samples is a dataframe.}

\item{sizes}{A numeric value for the size of the observed individual. Unused if samples is a dataframe.}

\item{weights}{(optional) The weight of an observation. For example the number of times a given age-size combination is observed. If weights are not provided all weights are set to 1. Unused if samples is a dataframe.}

\item{initpars}{A vector with initial parameters. Either an 5 element vector containing the elements and order as described below or  an elaborate vector with all parameters as created by \code{\link{makepoppars}}.
\itemize{
\item{the mean asymptotic size}
\item{the variance in asymptotic size}
\item{the mean length at the youngest age in the model}
\item{the variance in the size at the youngest age in the model}
\item{The growth rate of the Von Bertalanffy growth curve}
}}

\item{model}{A specific character expression specifying whether the asymptotic size is allowed to vary between timepoints.
\describe{
\item{\code{CONS}}{The asymptotic size does not vary between years}
\item{\code{VARY}}{The asymptotic size is allowed vary between years}
\item{\code{BOTH}}{Fit both models}
}}

\item{feedinggroups}{(optional) Indicate whether multiple feeding levels need to be estimated for different size- or agegroups.
\describe{
\item{\code{AGE}}{Feedinggroups are clustered by age}
\item{\code{SIZE}}{Feedingroups are clustered by size. note that the average size of a cohort is used and individual variation within a cohort is not considered when assigning the feedinggroup}
}}

\item{feedingbounds}{(Only if feedingroups are used) Vector with the boundaries of the feeding groups. The start of the fist feedinggroup and the end of the last feedinggroups do not need to be indicated.}

\item{logscale}{Boolean to indication whether the parameter space is explored on a log scale.}

\item{shrink}{Boolean to indicating whether the average size is allowed to shrink or should kept constant when the size exceeds the asymptotic size.}

\item{opts}{options passed to the \code{\link{nloptr}} function.}
}
\value{
list with detailes o the fitted models with a constant asymptotic size (CONS) and a varying asymptotic size (VARY).
For every fitted model the lists contain the following elements:
\describe{
\item{samples}{A dataframe with the observations used to fit the model.}
\item{globpars}{A vector with global parameters used for the modelfit containing
number of parameters,
number of timepoints,
lowest timepoint,
number of ages,
lowest age,
number of samples,
Shrinking allowance indication,
Log transformation indication,
feedinggroup indication
and number of feedinggroups}
\item{poppars}{Fitted model parameters in a vector as constructed by \code{\link{makepoppars}}.}
\item{asympsize}{Estimated mean and and variance of the asymptotic size per feedinggroup per year.}
\item{popstruct}{A list with the predicted mean and variance in size for every timepoint and age as constructed by \code{\link{predict_structure}}.}
\item{minLL}{The negative log likelyhood of the fitted model.}
\item{niter}{The number of iterations needed for the optimization.}
\item{initpars}{The initial parameters used for the optimization in a vector as constructed by \code{\link{makepoppars}}.}
\item{feedingbounds}{The boundaries of the feedinggroups used for the model.}
\item{opts}{The options used by \code{\link{nloptr}}.}
}
}
\description{
This function fits a timeseries of population structures to a dataset with size at age data.
To do so, it is assumed that the size at a given age at a given point in time is normally distributed and individuals follow a Von Bertalanffy growth curve.
The fitted asymptotic size is either constant or is allowed to vary between intervals.
}
\details{
The model is fitted by maximizing the log likelyhood using the \code{link{nloptr}} package.
}
