% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makepoppars.R
\name{makepoppars}
\alias{makepoppars}
\title{Make parameter vector defining the population stucture}
\usage{
makepoppars(
  fRlm_avg,
  fRlm_var,
  len0_avg,
  len0_var,
  rB,
  time0_avg = NULL,
  time0_var = NULL,
  ntime = NULL,
  nages = NULL,
  model = "BOTH",
  ngroups = 1
)
}
\arguments{
\item{fRlm_avg}{A single value for the mean asymptotic size,
or a vector with \code{ngroups} elements with a mean asymptotic size per feedinggroup
or if the model is \code{VARY} a vector with size \code{ntime-1} with the mean asymptotic size between every timestep
or if the model is \code{VARY} a vector with size \code{ngroups*(ntime-1)} with the mean asymptotic size per year for every feedinggroup (clustered per feedinggroup)}

\item{fRlm_var}{A single value for the variance in asymptotic size,
or a vector with \code{ngroups} elements with the variance in asymptotic size per feedinggroup
or if the model is \code{VARY} a vector with size \code{ntime-1} with the variance in asymptotic size between every timestep
or if the model is \code{VARY} a vector with size \code{ngroups*(ntime-1)} with the variance in asymptotic per year for every feedinggroup (clustered per feedinggroup)}

\item{len0_avg}{Vector of length \code{ntime} containing the average size at the youngest age in the model for every timepoint. If a single value is provided this is repeated for all timepoints.}

\item{len0_var}{Vector of length \code{ntime} containing the variance in size at the youngest age in the model for every timepoint. If a single value is provided this is repeaded for all timepoints.}

\item{rB}{single value for the Von Bertalanffy growth rate of the growth curve.}

\item{time0_avg}{(optional) vector of length \code{nage-1} containing the mean size at age at the first timepoint except for the mean size at the youngest age. If not given, this is calculated with the parameters for the first timepoint.}

\item{time0_var}{(optional) vector of length \code{nage-1} containing the variance in size at age at the first timepoint except for the variance in size at the youngest age.  If not given, this is calculated with the parameters for the first timepoint.}

\item{ntime}{number of timepoints in the model}

\item{nages}{number of ages in the model}

\item{model}{A specific character expression specifying whether the asymptotic size is allowed to vary between timepoints.
\describe{
\item{\code{CONS}}{The asymptotic size does not vary between years}
\item{\code{VARY}}{The asymptotic size is allowed vary between years}
\item{\code{BOTH}}{Fit both models}
}}

\item{ngroups}{Number of feeding groups in the model}
}
\value{
A vector with all model variables that are optimized by \link{fit_structure}.
If model is \code{CONS} the vector contains the following elements in order:
\itemize{
\item{average asymptotic size (1 element if model is \code{CONS}, \code{ngroups*(ntime-1)} elements if model is \code{VARY})}
\item{variance in asymptotic size (1 element if model is \code{CONS}, \code{ngroups*(ntime-1)} elements if model is \code{VARY})}
\item{average length at the youngest age in the model (\code{ntime} elements)}
\item{variance in length at the youngest age in the model (\code{ntime} elements)}
\item{average length at age at the first timepoint in the model except for the youngest age (\code{nages} elements)}
\item{variance in length at age at the first timepoint in the model except for the youngest age (\code{nages} elements)}
\item{The growth rate of the Von Bertalanffy growth curve (1 element)}
}
}
\description{
Make parameter vector defining the population stucture
}
