% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minLL_structure.R
\name{minLL_structure}
\alias{minLL_structure}
\title{Calculate the negative log-likelyhood for a model with given parameters for the given dataset.}
\usage{
minLL_structure(
  samples,
  ages = NA,
  sizes = NA,
  weights = NA,
  initpars,
  model = "BOTH",
  feedinggroups = NA,
  feedingbounds = NA,
  logscale = TRUE,
  shrink = TRUE
)
}
\arguments{
\item{samples}{A discrete (integer) value indicating the timepoint of the observation. Timepoints should have a constant interval and consecutive timepoints should have consecutive values. Or a dataset with 3 or 4 columns containing timepoints, ages, sizes and (optionally) weights as described.}

\item{ages}{A discrete (integer) value indicating the age of the observed individual. Ages should be on the same timescale as the timepoints. Unused if samples is a dataframe.}

\item{sizes}{A numeric value for the size of the observed individual. Unused if samples is a dataframe.}

\item{weights}{(optional) The weight of an observation. For example the number of times a given age-size combination is observed. If weights are not provided all weights are set to 1. Unused if samples is a dataframe.}

\item{initpars}{A vector with initial parameters. Either an 5 element vector containing the elements and order as described below or  an elaborate vector with all parameters as created by \code{\link{makepoppars}}.
\itemize{
\item{the mean asymptotic size}
\item{the variance in asymptotic size}
\item{the mean length at the youngest age in the model}
\item{the variance in the size at the youngest age in the model}
\item{The growth rate of the Von Bertalanffy growth curve}
}}

\item{model}{A specific character expression specifying whether the asymptotic size is allowed to vary between timepoints.
\describe{
\item{\code{CONS}}{The asymptotic size does not vary between years}
\item{\code{VARY}}{The asymptotic size is allowed vary between years}
\item{\code{BOTH}}{Fit both models}
}}

\item{feedinggroups}{(optional) Indicate whether multiple feeding levels need to be estimated for different size- or agegroups.
\describe{
\item{\code{AGE}}{Feedinggroups are clustered by age}
\item{\code{SIZE}}{Feedingroups are clustered by size. note that the average size of a cohort is used and individual variation within a cohort is not considered when assigning the feedinggroup}
}}

\item{feedingbounds}{(Only if feedingroups are used) Vector with the boundaries of the feeding groups. The start of the fist feedinggroup and the end of the last feedinggroups do not need to be indicated.}

\item{logscale}{Boolean to indication whether the parameter space is explored on a log scale.}

\item{shrink}{Boolean to indicating whether the average size is allowed to shrink or should kept constant when the size exceeds the asymptotic size.}
}
\value{
The negative log likelyhood of the model defined by the model parameters given the supplied samples.
}
\description{
Calculate the negative log-likelyhood for a model with given parameters for the given dataset.
}
