% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_structure.R
\name{predict_structure}
\alias{predict_structure}
\title{Calculate the population structure for a given parameterset}
\usage{
predict_structure(
  initpars,
  ntime,
  nages,
  mintime = 0,
  minage = 0,
  model = "BOTH",
  feedinggroups = NA,
  feedingbounds = NA,
  shrink = TRUE
)
}
\arguments{
\item{initpars}{A vector with initial parameters. Either an 5 element vector containing the elements and order as described below or  an elaborate vector with all parameters as created by \code{\link{makepoppars}}.
\itemize{
\item{the mean asymptotic size}
\item{the variance in asymptotic size}
\item{the mean length at the youngest age in the model}
\item{the variance in the size at the youngest age in the model}
\item{The growth rate of the Von Bertalanffy growth curve}
}}

\item{ntime}{number of timepoints in the model.}

\item{nages}{number of ages in the model.}

\item{mintime}{first timepoint in the model.}

\item{minage}{Lowest age in the model.}

\item{model}{A specific character expression specifying whether the asymptotic size is allowed to vary between timepoints.
\describe{
\item{\code{CONS}}{The asyptotic size does not vary between years}
\item{\code{VARY}}{The asyptotic size is allowed vary between years}
\item{\code{BOTH}}{Fit both models}
}}

\item{feedinggroups}{(optional) Indicate whether multiple feeding levels need to be estimated for different size- or agegroups.
\describe{
\item{\code{AGE}}{Feedinggroups are clustered by age}
\item{\code{SIZE}}{Feedingroups are clustered by size. note that the average size of a cohort is used and individual variation within a cohort is not considered when assigning the feedinggroup}
}}

\item{feedingbounds}{(Only if feedingroups are used) Vector with the boundaries of the feeding groups. The start of the fist feedinggroup and the end of the last feedinggroups do not need to be indicated.}

\item{shrink}{Boolean to indicating whether the average size is allowed to shrink or should kept constant when the size exceeds the asymptotic size.}
}
\value{
a list with the population structures predicted by the selected models. The population structures are saved as a matrix with the average and variance in asymptotic size in which the columns indicate the timepoints and the rows the different ages in the model.
}
\description{
Calculate the population structure for a given parameterset
}
